+++ tests/test_api.py
@@ -2,69 +2,17 @@
 
 
 import os
-import random
 import uuid
 
 import pytest
 
-from backlog import Backlog
-
-
-# @pytest.fixture denotes a function that is meant to be redefined:
-# pylint: disable=redefined-outer-name
-
-
-@pytest.fixture
-def backlog():
-    """Get a Backlog with at least 1 Entry."""
-    backlog_ = Backlog([
-        Backlog.Entry(title='test-1', priority=-100),
-        Backlog.Entry(title='test-2', priority=0),
-        Backlog.Entry(title='test-3', priority=1),
-        Backlog.Entry(title='test-4', priority=2),
-        Backlog.Entry(title='test-5', priority=100),
-    ])
-    random.shuffle(backlog_)
-    return backlog_
-
-
-@pytest.fixture
-def backlog_entry(backlog):
-    """Get a Backlog and an Entry it contains."""
-    return backlog, random.choice(backlog)
-
-
-@pytest.fixture
-def entry():
-    """Get a Backlog.Entry."""
-    return Backlog.Entry(
-        title='test-foo',
-        priority=10,
-    )
-
-
-@pytest.fixture
-def path(tmpdir):
-    """Get the path of a writable file."""
-    path_ = str(tmpdir.join('{0}.json'.format(uuid.uuid4())))
-    yield path_
-    try:
-        os.remove(path_)
-    except FileNotFoundError:
-        pass
-
-
-@pytest.fixture
-def saved_backlog(backlog, path):
-    """Get a Backlog and a path to a file where it has been saved."""
-    backlog.save(path=path)
-    return backlog, path
+import backlog.api as api
 
 
 def test_backlog_load(saved_backlog):
     """Verify that a saved Backlog is equivalent when re-loaded."""
     backlog, path = saved_backlog
-    assert backlog == Backlog().load(path)
+    assert backlog == api.Backlog().load(path)
 
 
 def test_backlog_random(backlog):
@@ -92,7 +40,7 @@ def test_backlog_search_invert(backlog):
 
 def test_entry_equality_duplicate(entry):
     """Backlog.Entry objects are equal if they have the same attributes."""
-    assert entry == Backlog.Entry(
+    assert entry == api.Backlog.Entry(
         note=entry.note,
         title=entry.title,
         priority=entry.priority,
@@ -122,5 +70,5 @@ def test_entry_summary_is_str(entry):
 def test_repr(request, fixture):
     """Ensure that __repr__ is a valid serialization."""
     value = request.getfixturevalue(fixture)
-    locals()['Entry'] = Backlog.Entry  # Entry is an inner class.
+    locals()['Entry'] = api.Backlog.Entry  # Entry is an inner class.
     assert value == eval(repr(value))  # pylint: disable=eval-used
